requirements.txt
flask
pydub
openai-whisper
fpdf